Rename Query => Queries
