+++ src/core/Queries.ts
