+++ src/index.ts
@@ -9,7 +9,8 @@ export * from './core/expr/Conditions.ts'
 export * from './core/expr/Include.ts'
 export * from './core/Index.ts'
 export * from './core/Internal.ts'
-export * from './core/Query.ts'
+export * from './core/Queries.ts'
+export * from './core/query/CTE.ts'
 export * from './core/query/Delete.ts'
 export * from './core/query/Insert.ts'
 export * from './core/query/Select.ts'
